+++ test/unit/libraries/ConstantProductMath.t.sol
@@ -12,28 +12,28 @@ contract ConstantProductMathTest is BaseTest {
         assertEq(lSpotPrice, 1e18);
         assertEq(lLogSpotPrice, LogCompression.toLowResLog(1e18));
 
-        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint112).max, type(uint112).max);
+        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint104).max, type(uint104).max);
         assertEq(lSpotPrice, 1e18);
         assertEq(lLogSpotPrice, LogCompression.toLowResLog(1e18));
 
-        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint112).max / 2, type(uint112).max);
+        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint104).max / 2, type(uint104).max);
         assertEq(lSpotPrice, 2e18 + 1);
         assertEq(lLogSpotPrice, LogCompression.toLowResLog(2e18));
 
-        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint112).max, type(uint112).max / 2);
+        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint104).max, type(uint104).max / 2);
         assertEq(lSpotPrice, 0.5e18);
         assertEq(lLogSpotPrice, LogCompression.toLowResLog(0.5e18));
 
-        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint112).max / 10, type(uint112).max);
+        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint104).max / 10, type(uint104).max);
         assertEq(lSpotPrice, 10e18 + 1);
         assertEq(lLogSpotPrice, LogCompression.toLowResLog(10e18));
 
-        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint112).max, type(uint112).max / 10);
+        (lSpotPrice, lLogSpotPrice) = ConstantProductOracleMath.calcLogPrice(type(uint104).max, type(uint104).max / 10);
         assertEq(lSpotPrice, 0.1e18);
         assertEq(lLogSpotPrice, LogCompression.toLowResLog(0.1e18));
 
         (lSpotPrice, lLogSpotPrice) =
-            ConstantProductOracleMath.calcLogPrice(type(uint112).max, type(uint112).max / 1e17);
+            ConstantProductOracleMath.calcLogPrice(type(uint104).max, type(uint104).max / 1e17);
         assertEq(lSpotPrice, 1e18 / 1e17);
         assertApproxEqRel(
             lLogSpotPrice,
@@ -43,7 +43,7 @@ contract ConstantProductMathTest is BaseTest {
         );
 
         (lSpotPrice, lLogSpotPrice) =
-            ConstantProductOracleMath.calcLogPrice(type(uint112).max / 1e17, type(uint112).max);
+            ConstantProductOracleMath.calcLogPrice(type(uint104).max / 1e17, type(uint104).max);
         // this small discrepancy occurs because we use divWadUp
         assertApproxEqRel(lSpotPrice, 1e18 * 1e17, 0.00000000000000001e18);
         assertApproxEqRel(
@@ -54,12 +54,12 @@ contract ConstantProductMathTest is BaseTest {
         );
 
         (lSpotPrice, lLogSpotPrice) =
-            ConstantProductOracleMath.calcLogPrice(type(uint112).max, type(uint112).max / 1e18);
+            ConstantProductOracleMath.calcLogPrice(type(uint104).max, type(uint104).max / 1e18);
         assertEq(lSpotPrice, 1e18 / 1e18);
         assertApproxEqRel(lLogSpotPrice, LogCompression.toLowResLog(1e18 / 1e18), 0.1e18);
 
         (lSpotPrice, lLogSpotPrice) =
-            ConstantProductOracleMath.calcLogPrice(type(uint112).max / 1e18, type(uint112).max);
+            ConstantProductOracleMath.calcLogPrice(type(uint104).max / 1e18, type(uint104).max);
         // this small discrepancy occurs because we use divWadUp
         assertApproxEqRel(lSpotPrice, 1e18 * 1e18, 0.000000000000001e18);
         assertApproxEqRel(lLogSpotPrice, LogCompression.toLowResLog(1e18 * 1e18), 0.1e18);
@@ -67,8 +67,8 @@ contract ConstantProductMathTest is BaseTest {
 
     function testCalcLogPrice_ReturnsOneWeiWhenPriceDiffGreaterThan1e18(uint256 aReserve0, uint256 aReserve1) public {
         // arrange
-        uint256 lReserve1 = bound(aReserve0, 1, type(uint112).max / 1e18);
-        uint256 lReserve0 = bound(aReserve1, lReserve1 * 1e18, type(uint112).max);
+        uint256 lReserve1 = bound(aReserve0, 1, type(uint104).max / 1e18);
+        uint256 lReserve0 = bound(aReserve1, lReserve1 * 1e18, type(uint104).max);
 
         // act
         (, int112 lLogPrice) = ConstantProductOracleMath.calcLogPrice(lReserve0, lReserve1);
